Feat: 지역,대학 선택 기능 완료
